+++ src/options.cpp
@@ -46,7 +46,7 @@ bool Options::parse(int argc, char **argv)
 	std::stringstream base;
 	if (layers.size() > 0)
 		base << prefix << "-" << layers[0];
-	for (int l = 1; l < layers.size(); l++)
+	for (size_t l = 1; l < layers.size(); l++)
 		base << "x" << layers[l];
 	basename = base.str();
